Split PdbCondition into classes for each behaviour.
